requirements.txt
anthropic>=0.40.0
sqlalchemy>=2.0
lightrag-hku>=1.0.0
pyyaml>=6.0
json-repair>=0.30.0
fastapi>=0.115.0
uvicorn>=0.30.6
